Install CImg.h as gmic_CImg.h.
As discussed with upstream we ship gmic's CImg.h as
gmic_CImg.h as it is usually a pre-release / unfinished
version, compared to the one released and shipped in
the cimg-dev package. People who want to use gmic's
features and link with libglib need the version gmic
was built with - and building gmic with the released
version is not possible.

Closes: #801162