Link all binaries dynamically
Switch the build of all the binaries to dynamically link to libgmic,
instead of building their own copies of gmic; this drastically reduces
the build time, and avoids different build configurations for the gmic
build copies in the binaries.

Add a patch for zart so it is possible to specify the right path to the
build gmic libraries.

As result, tighten the libgmic1 dependencies everywhere.